Download File Progress
The progress is reporting the number of downloaded pieces, which can be larger than 100.
To download the progress in percentage (ranging from 0 to 100) devide size downloaded size by total file size and multiply by 100 (for percentage). Use Math.Floor to not report a progress of 100 when a byte is still missing.